+++ RELEASE-NOTES.md
@@ -1,7 +1,13 @@
 
-# 2.0.2 (PLANNED)
+# 2.0.3 (PLANNED)
 
 TBD
+
+# 2.0.2
+
+* Add more standard error codes (such as 429 Too Many Requests).
+* Rename standard error code ``UNSUPPORTED`` to ``NOT_IMPLEMENTED`` to be consistent with the corresponding HTTP status.
+* Remove dependency on servlet-api.
 
 # 2.0.1